Add unofficial completions guide for fish shell
Hello; such a great project! I'm just learning k8s, and I like my CLI UI to be clean and readable.

As I've been using fish for a while now, I was sad to discover I can't use `kubecolor` with proper completions with fish. Completions are necessary to me for learning, so I tried to make it work.

As I'm no expert when it comes to hacking `fish`, I just took @evanlucas (thx ! 🙏 ) completions, and replaced all `kubectl` with `kubecolor`. It's ugly, but it works.

@evanlucas, perhaps you could guide me how to do it in a cleaner way. I know little about `go`, and seems you're using it to generate these completions. Worst case, I could just set up a github action that periodically fetches your `kubectl` completions and builds `kubecolor` ones from that.